Move review classes in their own module/package.
