+++ asreview/__init__.py
@@ -1,6 +1,7 @@
-from asreview.base import ReviewSimulate, ReviewOracle
 from asreview.readers import read_csv, read_data, read_ris, ASReviewData
-from asreview.review import review, review_oracle, review_simulate, get_reviewer
+from asreview.review import (ReviewSimulate, ReviewOracle, MinimalReview,
+                             review, review_oracle, review_simulate,
+                             get_reviewer)
 from asreview.utils import text_to_features
 from asreview.models.embedding import load_embedding, sample_embedding
 from asreview.logging import Logger, read_log, read_logs_from_dir
