+++ asreview/review/simulate.py
@@ -1,0 +1,61 @@
+from asreview.review import BaseReview
+from asreview.base import _merge_prior_knowledge
+from asreview.init_sampling import sample_prior_knowledge
+
+
+class ReviewSimulate(BaseReview):
+    """ Automated Systematic Review in simulation mode. """
+
+    def __init__(self,
+                 X,
+                 y,
+                 n_prior_included=None,
+                 n_prior_excluded=None,
+                 *args, **kwargs):
+        super(ReviewSimulate, self).__init__(
+            X, y, *args, **kwargs)
+
+        self.n_prior_included = n_prior_included
+        self.n_prior_excluded = n_prior_excluded
+
+    def _prior_knowledge(self):
+        if self.prior_included and self.prior_excluded:
+            prior_indices, prior_labels = _merge_prior_knowledge(
+                self.prior_included,
+                self.prior_excluded
+            )
+
+            return prior_indices, prior_labels
+
+        elif self.n_prior_included and self.n_prior_excluded:
+
+            # Create the prior knowledge
+            init_ind = sample_prior_knowledge(
+                self.y,
+                n_prior_included=self.n_prior_included,
+                n_prior_excluded=self.n_prior_excluded,
+                random_state=None  # TODO
+            )
+
+            return init_ind, self.y[init_ind, ]
+        else:
+            raise ValueError(
+                "provide both prior_included and prior_excluded, "
+                "or n_prior_included and n_prior_excluded"
+            )
+
+    def _get_labels(self, ind):
+        """ Get the labels directly from memory.
+
+        Arguments
+        ---------
+        ind: list, np.array
+            A list with indices
+
+        Returns
+        -------
+        list, np.array
+            The corresponding true labels for each indice.
+        """
+
+        return self.y[ind, ]
